refactor: merge signups into user router
