--- app/api/signup/__init__.py
@@ -1,1 +1,0 @@
-"""Common app that will be used in project."""
