--- app/api/signup/views.py
@@ -1,10 +1,0 @@
-from fastapi import APIRouter
-from .schemas import Signup
-
-
-router = APIRouter()
-
-
-@router.post("/", response_model= Signup)
-async def create_signup(signup: Signup):
-    return signup
